Split handling of scopes and functions in TypeInference
Summary:
In preparation for moving scopes out of functions, treat Variables
separately from functions in TypeInference. They are treated as their
own top level construct during partitioning, and are added to the same
set as any function that uses them.

This has the same effect, since functions that share data through
Variables will still end up in the same set.

Note that it is important to do this on Variables rather than
VariableScopes, because a union of all of the users of a VariableScopes
would end up including all functions in the module. This happens
because all VariableScopes are chained, and therefore indirectly tied
to the top level scope.

Reviewed By: avp

Differential Revision: D54763372

fbshipit-source-id: 481d4f82f2bd50ee538338a7e068808920855a03