Clean up modifications to DebuggerAPI when destroying CDPAgent
Summary:
When destroying `DebuggerDomainAgent` the code doesn't clean up every
modification it made to `DebuggerAPI`. Added a `cleanUp()` function to
properly unregister everything. Chose to name the helper function
`cleanUp()` instead of `disable()` in the end. This is so that there's
no confusion in the future if we do decide to persist agent state.

Reviewed By: motiz88

Differential Revision: D54898267

fbshipit-source-id: 7d9baea7221a1f69828764938fb5de0737419f6b