MiniReact: Lower unsupported features only (loose mode)
Summary:
Use the `assumptions` config to specify loose mode for classes and
properties, and don't run the entire preset, just the features we don't
support.

This brings performance almost in line with webpack.
The only thing that remains is wrapping untyped code in an IIFE,
which webpack does but `flow-bundler` does not.

Rename from `-es5` to `-lowered` because this is no longer ES5 code.

Reviewed By: pieterv

Differential Revision: D55028581

fbshipit-source-id: 45a0c6b49d17887b802720a2acff0eb2dbebe9bf