A Range trait for accessing .range on nodes and enums
Summary:
Previously we added range() function for all ESTreeNode for accessing .range. But now I need to access `.range` on a enum type `FlowTypeAnnotation`, which isn't techinically a ESTreeNode.
This diff adds a separate `Range` trait for getting `range`, to make it possible to access `.range` on both enums and nodes

Reviewed By: evanyeung

Differential Revision: D55040872

fbshipit-source-id: 03896e58bede3a407c45ea8d5124c15b8a948d14