Avoid handling AsyncBreak for single stepping Ret
Summary:
This is another root cause of the debugger becoming stuck if `Ret`
OpCode is replaced with a `Debugger` OpCode. After stopping at the
`Debugger` OpCode, we try to single step the `Ret` OpCode by calling
`stepFunction()`. Then while handling the `Ret`, the code checks for
AsyncBreak and calls `DISPATCH`, which will return from
`interpretFunction` in single-step mode. This causes the single step to
not actually process any OpCode, so the debugger becomes stuck and never
move past the `Ret` overridden by `Debugger` OpCode.

Reviewed By: avp

Differential Revision: D55092318

fbshipit-source-id: 3c44917b3b07d4e6f41e58bc71fb561b54b09a33