Get the correct OpCode in vm::Debugger
Summary:
The debugger can become stuck if a breakpoint is installed on a `Ret`
OpCode. One of the problems in that scenario is that the debugger code
doesn't know it should perform a StepOut. This is because
`CodeBlock::getOpCode()` doesn't look at the underlying OpCode.

Fixing it by adding a `getRealOpCode()` function, which should be used
whenever we're basing some decision on an OpCode. We should always be
making decisions based on the original OpCode.

Reviewed By: avp

Differential Revision: D55047861

fbshipit-source-id: 5f10b1a5ab419daad2cc8759ef3b36eb07e78b6c