Send correct Debugger.paused reason for StepFinish
Summary:
When stepping through code, the `Debugger.paused` reason is "step" in
Chrome. Updating the code to match. At the same time deduped the
`sendPausedNotificationToClient()` and
`sendPauseOnExceptionNotificationToClient()`. The only difference
between the two was the reason field.

Aside from fixing the reason to match Chrome, this allows us to verify
that a pause is due to StepFinish.

Reviewed By: motiz88

Differential Revision: D55216788

fbshipit-source-id: 90ef1eb637e4e5758755d7b948a8a4dbbc07b31f