Fix debugger not stopping after step over a throw
Summary:
Code duplication is hurting us by introducing bugs from discrepancy.
The code in `runDebugger` does not call `breakpointExceptionHandler()`
and so stepping over a throw didn't pause again. The code in
`debuggerLoop` does call `breakpointExceptionHandler()`.

This diff doesn't try to solve the code duplication. Will treat that as
a future TODO because it's not quite trivial.

Reviewed By: avp

Differential Revision: D55224088

fbshipit-source-id: 61217f2bba67abc362b2cd3730d808e211a2ee3d