use ConsecutiveStringStorage to dedup serialized literals
Summary:
Serialized literals de-duping was quadratic with the total size of the
literals because it performed a linear search when adding every literal.

We can use ConsecutiveStringStorage (with small modifications) instead
to dedup the literals with much better performance. To that end, we need
an additional pass over the LIR to collect all literals, dedup them, and
remember their dedupped offsets.

Original Git: 221ce21
Original Revision: D18056232

The only real change from the original diff is in relation to writing
the outputs of `LiteralBufferBuilder::generate`. Previously, it wrote
the literal buffers directly into BMGen via
`bmGen_.initializeSerializedLiterals`. However we want this in SH as
well, so `generate` returns the buffers, and then each backend will
write the result to the correct place. `LiteralBufferBuilder` must also
now take function pointers to lookup strings and symbols, because
`LiteralGenerator`'s constructor needs that information. This change to
`LiteralGenerator` was only made in SH. Also, `LiteralBufferBuilder` is
extracted into a separate file so both backends can use it.

Reviewed By: avp

Differential Revision: D53967862

fbshipit-source-id: 5bd46532a98d178e16a2d21a6e6c0fc380badd56