Include StackTrace for consoleAPI notification
Summary:
Adds `StackTrace` to `ConsoleMessage`. Due to the way console API is
setup in React Native, I expect RN will need to manipulate the list of
`CallFrame`s to remove frames of RN internal code in order to show
where the actual console API call is. So instead of `CDPDebugAPI` doing
it for the integrator, leaving it up to the integrator to provide the
`StackTrace`.

Then include the `StackTrace` when sending
`ConsoleAPICalledNotification`.

Reviewed By: mattbfb

Differential Revision: D55274020

fbshipit-source-id: 677c99653deb2f4ff610f5289c0332d6476c7d39