Move the rest of generateBytecodeModule into BytecodeGenerator
Summary:
TraverseLiteralStrings was a collection of multiple very simple
functions which took very simple `std::functions` and only had one
callsite.

All of this can be inlined and put into a `BytecodeGenerator` function
for collecting the strings.
This lets us move the other code which relies on the results of
collecting the strings into the `BytecodeGenerator`,
and just use temporary state inside `BytecodeGenerator`.

All this is done in a single diff because splitting it was slightly more
painful and we'd end up here anyway.

Reviewed By: neildhar

Differential Revision: D55451150

fbshipit-source-id: 41703eaa199f6c4b192c2e50cfbb97a9e7a2241b