Modify ES6 Method names instead of removing them (#1363)
Summary:
The method names of ES6 classes were being stripped out to prevent name conflicts when there's both a function and method with identical names. However, when connecting the Debugger or Profiler, this causes all class methods to show up with the name `anonymous` instead of its actual function name.

This change modifies the ES6 method names to prefix and suffix them with '#' instead of removing it, so that the debugger and profiler give more detailed information, while still preventing the name conflicts.

Before:
![Screenshot 2024-03-28 at 2 46 57 PM](https://github.com/facebook/hermes/assets/34820577/ee8cd189-c5d6-479b-973b-22099088bc6d)

After:
![Screenshot 2024-03-28 at 2 42 36 PM](https://github.com/facebook/hermes/assets/34820577/6aea903d-caf0-4a97-96a5-39ca1897a973)

Pull Request resolved: https://github.com/facebook/hermes/pull/1363

Test Plan: .

Reviewed By: avp

Differential Revision: D55511103

Pulled By: tmikov

fbshipit-source-id: fb73b8bc87fe27d372a91969ef9c7088d8e8c702