Add exception handling with Result
Summary:
## Context

Currently, if there is a missing wrapper type or mismatch between the wrap with hermes, the crates panics and it bubbles up. It would terminate the relay compiler and doesn't show what source is causing the error. We wouldn't want that espicially during the rollout phase of new resolver parser.

## This diff
Add DiagnosticsResult to all `convert` functions, so we can later make all panics into a proper error

Reviewed By: captbaritone

Differential Revision: D55457401

fbshipit-source-id: db31d07cc3e662d82d3614a9f80413c95f4290ca