Track parent scope in VariableScope
Summary:
Keep track of the parent scope for each VariableScope. This will allow
us to determine scope depth and parents without needing to encode it in
the instructions.

Each VariableScope now participates in two intrusive lists. The first is the
existing owning list in Module. The second is a non-owning list of
children of its parent VariableScope. The child list is necessary so that
we can update the parents of the children when a scope is eliminated.

Reviewed By: tmikov

Differential Revision: D54473919

fbshipit-source-id: 423e660cc0269765a8ce8a7a84b0194b120519b8