Expose HERMES_DEBUGGER_ENABLED via HermesInternal.getRuntimeProperties
Summary:
Expose the build flag `HERMES_DEBUGGER_ENABLED` at runtime via the
"Debugger Enabled" prop on `HermesInternal.getRuntimeProperties`.

`HERMES_DEBUGGER_ENABLED` is typically set for `dev` and `dbg`
builds of Hermes, as well as profile builds of apps, but it may be
specifically overridden so I don't know of a reliable way to infer
it when diagnosing a prebuilt app.

Reviewed By: avp

Differential Revision: D55871440

fbshipit-source-id: a0ddb9e525f0f3efc4b96a5a23745dcb8a68b547