Lower inner generator function into a switch statement
Summary:
Transform `SaveAndYield`s in a generator function into a regular
function with a switch statement that jumps to the correct block. This
also requires a transformation of the try instructions of the inner
function, since it is illegal to jump from the switch directly into a
BB that was initially wrapped in a try.

Reviewed By: avp

Differential Revision: D53096972

fbshipit-source-id: d038fb0dec24c49f11da61881c99e36b0f83ca74