Use RecursiveVisitor to pass Type constraint to expressions
Summary:
Pass `nullptr` constraint to all expressions right now.
The actual propagation will be in a later diff.

Add a comment explaining what the parameter is for and how it's used.
Passing a `constraint` allows the child to perform any checks it wants,
but the caller that passed the constraint doesn't make assumptions about
the actual type assigned to the node.

Reviewed By: neildhar

Differential Revision: D53767977

fbshipit-source-id: cb1aa6e4ae993be124fde92f57cbc56e8c78a52b