Add extra info to RecursiveVisitor
Summary:
The dispatcher takes optional type parameters ExtraT,
which can be passed as the last arguments to the dispatcher.
This allows the visitor to pass information to the recursive calls.
The dispatcher will call the appropriate overload of the visitor's
visit() method.

NOTE: visit() methods that take ExtraT must take a parent
argument as well, to avoid confusion around whether the parent pointer is
extra info or not.

visitChildren can also be called with ExtraT, which will be propagated to
every child node. Similarly, visitESTreeNodeList can be called with
ExtraT, which will be propagated to every element of the list.

It's convenient to implement this as a variadic,
which results in no extra code if ExtraT is not used but also allows for
full flexibility to add as many extra info as the caller would like to
pass. We can also extend this in the future to make the ExtraT optional
to pass to visit and default-construct it, but that wasn't used yet so
we can add it later.

Reviewed By: tmikov

Differential Revision: D53767975

fbshipit-source-id: 85db9d95d9a1aa587b09f96299dc48609a523a92