Infer types of arrays/tuples using the constraint mechanism
Summary:
This only handles the currently supported node types, but it does it
using the constraint parameter instead of the parent parameter, meaning
it works for nested tuples/arrays.

The `constraint` allows for errors to be emitted if the typechecked type
isn't able to match the constraint, but that's not required of the
callee that receives the constraint.

Reviewed By: neildhar

Differential Revision: D53821078

fbshipit-source-id: dc041f555d3b1a50f777fe4eb3c5fa14571f464e