Fix invalid iterator access in splitBasicBlock
Summary:
`Instruction::moveBefore` needs to be able to `getParent` on its
operand. However, `splitBasicBlock` was calling it with an end
iterator, which is invalid.

To avoid running into this in the future, remove the new iterator
version of `moveBefore` and add a comment.

`splitBasicBlock` now doesn't operate on the terminator created by the
callback. Setting the location is instead entirely the responsibility
of the callback. This simplifies things, since it is easier to just
set the location once in the builder when creating the instruction,
instead of moving it afterwards. (which is further complicated by the
fact that the block is being split)

Reviewed By: tmikov

Differential Revision: D56045320

fbshipit-source-id: f76ea2ef2a05de7452f200e9aae62bf535a9c0e0