Back out "Add captureStackTrace to DebuggerAPI"
Summary:
Backing this out because I realized that I actually shouldn't be using
`getCurrentInterpreterLocation()`. That function traverses the stack
frames to find the first place where a CodeBlock can be found. This
causes the `captureStackTrace()` to display wrong information for the
top frame.

Fixing it in the subsequent diff.

Original commit changeset: 10fde18ff951

Original Phabricator Diff: D55273833

Reviewed By: mattbfb

Differential Revision: D55757946

fbshipit-source-id: e7cfeb707f7d3ee53535d6b6789d58863ec3131f