Add captureStackTrace to DebuggerAPI take 2
Summary:
Re-introduce this change after backing out the previous implementation
to fix a bug.

Parallel of the `V8Debugger::captureStackTrace()` used by Chrome to
report their call stack for console APIs.

The DebuggerAPI `captureStackTrace()` handles the situation when the
interpreter loop isn't running, so as long as the caller is on the
runtime thread, the function can be used.

Reviewed By: avp

Differential Revision: D55757947

fbshipit-source-id: d34c6a36810e2eb23443f3d0fab17b99b8b18bdb