Add Hermes CDP CLI
Summary:
Add an initial version of the `hcdp` tool, which is a utility
to run Hermes CDP from the CLI.

This tool is a successor to the "hermes chrome debug tool"
from previous versions of Hermes CDP.

Like the the hermes chrome debug tool, this currently
carries out debugging so long as the script is running,
so longer-running scripts are ideal (e.g. scripts with a
loop, or a `debugger;` statement).

As the tool expands, we may move to have a more
sophisticated build process (e.g. typescript compilation,
multiple files, bundling, etc.), but it's kept simple for now.

Reviewed By: dannysu

Differential Revision: D54383479

fbshipit-source-id: 0148ad48cac3178a411a9ca563e83556ec72dd11