Move destruction of RuntimeModule to the mutator (#1374)
Summary:

There is race condition between removing a RuntimeModule from the
Runtime in the background thread (triggered from Domain finalizer)
and adding a new RuntimeModule in the mutator.

After further investigation, we found that there is also race 
conditions between Debugger::willUnloadModule() called in the 
destructor of RuntimeModule and other Debugger actions.

This diff moves the destruction of RuntimeModule (and its removal from
the list in Runtime) to the time of marking weak roots in Runtime.

Reviewed By: neildhar

Differential Revision: D56094070