Get basic block profiling to compile.
Summary:
This diff gets basic block profiling to work again.  (Which also enables function-level profiling by # of bytecode instructions executed, and counting of frequency of bytecodes.)

To do this, I removed all the stuff related to tracking of inline caching hit rates.  This wasn't compiling when I enabled the flag; I think some of the code was originally in the interpreter loop, but got migrated out to a helper function, where it wouldn't compile because it used local variables of the interpreter loop.  Inline caching profiling seems separate from BB profiling.  If we want to re-enable it, we have the Hermes code to use as a guide.  Or we can find this diff and reinstate it.

Additionally:

* The -basic-block-profiling option just wrote the JSON file to stderr.  Which isn't ideal in cases where you have other things, like compiler warnings, being written to stderr as well.  So I added a -profiling-out option to specify the output file for this data.  This exposed the fact that CompilerDriver.cpp had a pretty useful abstraction, OutputStream, that was private to the .cpp file.  I moved that to it's own .h file in the support directory, to use it in ConsoleHost.

FWIW: I don't have high confidence in the "chain of custody" of line/column numbers -- the source locations given for the starts of basic blocks in the basic block profiles seem a little questionable.  But I haven't looked into it.

Reviewed By: avp

Differential Revision: D55592754

fbshipit-source-id: de91f639f9b20d67b7b3d125ad282dc20da48fa3