Factor out and simplify resolve scope optimisation
Summary:
Factor out the InstSimplify optimisation for `ResolveScopeInst` into a
separate function in IRUtils. This will be used to elide `ResolveScope`
in IRGen in upcoming diffs.

While doing this, also make a slight simplification to this loop, by
removing the requirement that the obtained scope is a `BaseScopeInst`.
This is no longer necessary since scopes now track their parents, so
we can compute the parent even when it isn't a `BaseScopeInst`. This
should allow for slightly better results in the very rare case that a
scope's parent is not a `BaseScopeInst`.

Reviewed By: avp

Differential Revision: D56259271

fbshipit-source-id: 9ed9e85110ffd40745df841660e61711ef61133d