Rename functionScope to curScope in IRGen
Summary:
Rename functionScope to curScope now that we may have multiple scopes
in a function.

Eventually, this will need to be renamed to curEnv, but I'm avoiding
directly going to that to keep some semblance of consistency for now.
Once in flight changes that touch scopes are all landed, we can rename
everything at once.

Reviewed By: avp

Differential Revision: D56026169

fbshipit-source-id: 4a20011b229535797b86227ac4365673ec7cbf60