Fix atob not accepting encoded strings without padding
Summary:
The forgiving-base64 decode algorithm
(https://infra.spec.whatwg.org/#forgiving-base64) gives specific
instructions for if remainder is 0 and if remainder is 1.

The only error case is when the remainder is 1:
> If data’s code point length divides by 4 leaving a remainder of 1,
then return failure.

When the remainder is 2 or 3, it should still be considered valid
base64 data.

Fixing `base64DecodeOutputLength()` so we're not returning error
incorrectly.

Reviewed By: avp

Differential Revision: D56480630

fbshipit-source-id: ef500ddd6ed1a128150a79f6d7d8ae4ef8be4f2d