Add Set fast paths in Set constructor
Summary:
Reading from `Set` when the iterator is the default iterator can be made
faster by just iterating the entries directly.

Improves speed of a microbenchmark that just clones a 1000-element
Set ~6x.

NOTE: This can be further improved by bypassing the extra
hashing/expansion but that's going to require more effort. This is just
a first version of the fast path.

Reviewed By: fbmal7

Differential Revision: D56324024

fbshipit-source-id: 3d39ee8ec9205e23519a189e4821de4e423ff39a