Get the correct OpCode in vm::Debugger
Summary:
Original Author: dsc@meta.com
Original Git: f09c55368d08ab21d515a6ad30271dd10710d573
Original Reviewed By: avp
Original Revision: D55047861

The debugger can become stuck if a breakpoint is installed on a `Ret`
OpCode. One of the problems in that scenario is that the debugger code
doesn't know it should perform a StepOut. This is because
`CodeBlock::getOpCode()` doesn't look at the underlying OpCode.

Fixing it by adding a `getRealOpCode()` function, which should be used
whenever we're basing some decision on an OpCode. We should always be
making decisions based on the original OpCode.

Reviewed By: mattbfb

Differential Revision: D56482764

fbshipit-source-id: 2ff1fca2f32e339fd0a2b4cc1c1a4afb4c2f07f4