Narrow type of HermesInternal.concat call to string
Summary:
When we lower StringConcat to a call to `HermesInternal.concat`, the
type of the call is set to `any`. This means that after the pass, the
value has a wider type than it did before it. This in turn causes
verification problems, since instructions expect the narrower type.

To fix this, narrow the type of the call back to string.

Reviewed By: avp

Differential Revision: D56617407

fbshipit-source-id: 9493d83a772a19123f3a79324e91f00e977663e4