Fix processing of Terminators in moveCrossingValuesToOuter
Summary:
Terminators must be the last instruction in a BB. Therefore, the code we
had in the generator pass for dealing with AllocStack was incorrect.
After instructions that write to an AllocStack, we need to
propagate the value of the dummy AllocStack to the outer variable. We
were blindly doing this for all instructions. But, there are some
terminators that take in an AllocStack. Instead, we must do the
write-back of the dummy AllocStack in the successors of the terminator.

It would not be correct to emit the write-back IR to the successors of
the terminator as-is. This wouldn't be correct for instance in looping
BBs. So we create a new layer of BBs. Each existing successor of the
terminator is replaced by a new BB. This BB performs the write-back, and
then branches to the original successor.

Reviewed By: avp

Differential Revision: D56579156

fbshipit-source-id: ca5749f1e9ae725782eb6a82e436afb6592d56e3