NULL Guard Page
Summary:
allow-large-files

Set the first 16 KiB (default) of the WASM VM's
memory to be a guard page. 16 KiB is the page
size for ARM64, so defaulting to use a multiple
of this regardless of underlying system.

The size of the guard page/region is configurable
by setting the `GLOBAL_BASE` linker variable
used by [Emscripten](https://emscripten.org/).
This value is used to determine the offset the
data region will start at. Furthermore, this value
is pulled via a CMake function and passed in as
a compile time definition.

Reviewed By: neildhar

Differential Revision: D56068127

fbshipit-source-id: be45104d92b28f644c7586a75d9b34fd172c0423