Stack Clash Guard Page
Summary:
Enable support for guard page between top of data
section and bottom of WASM stack. This provides a
near 0 CPU cost for detecting stack overflowing into
the data section and causing data corruption.

When enabled, the last 16 KiB (aligned to 16 KiB) of
the stack will be used as a guard page, meaning that
a worst case of 32 KiB - 1 of stack space will be
claimed for the guard page, leaving a default of
224 KiB of stack space for WASM. The 16 KiB
requirement is due to page sizes on ARM64 being
16 KiB.

Reviewed By: neildhar

Differential Revision: D56192935

fbshipit-source-id: 96beb892e324884322678d16d790ff42b3c3ca71