Fix indentation bug and strict mode bug in testsuite.py
Summary:
The wrong indentation causes this script only run bytecode generated
with -O0, but we intended to run both to catch bugs in optimizations.

Another issue is that `strictEnabled` is not correctly assigned, it's
currently a loop variable so the value is always from last iteration.
As a result, the script always run no strict mode in case that it
should run both.

Reviewed By: avp

Differential Revision: D56430910

fbshipit-source-id: 3dd99028c648b2bf7f9f3b2f5574a140803b0ad5