Improve IR gen for SwitchBuilder
Summary:
The SwitchBuilder in LowerGeneratorFunction has a convenient assumption
it can make: every time the switch is executed, the switch input is
guaranteed to match one of the cases. This fact enables the generation
of more efficient IR. Specifically:

- Switches with 1 input can be expressed as an unconditional branch.
- For N cases, we only need to perform N-1 comparisons. If we haven't
found a match looking at the first N-1 cases , we know we can branch to
the last one. We use the default block mechanism that already exists to
do this.

Reviewed By: avp

Differential Revision: D57061568

fbshipit-source-id: 5b0e78ce9684c0812b31390edf14f1a6c0e6ac0c