Use PointerValue for ObjectID
Summary:
Use `PointerValue*` as a source of `SynthTrace::ObjectID` instead of
`UniqueID` from `Runtime` which is based on GC's unique ID system. This
will reflect better about the life time of each `jsi::Pointer` in the
record so that we can emulate the closer behavior to the original when
we replay the record.

Note that, the non-deterministic hooks for WeakRef needs to be changed
because the existing logic was relying on the fact that the ObjectID
for the referenced JSObject won't change between WeakRef creation and
deref time, which is no longer true.

Reviewed By: neildhar

Differential Revision: D57133457