Omit "this" from local scopes
Summary:
The stack frame object has [a dedicated spot for
`this` values](https://chromedevtools.github.io/devtools-protocol/tot/Debugger/#type-CallFrame). We report the `this` value via this
mechanism, so also including `this` in the scope
properties produced duplicate `this` values.

Remove the `this` from scope properties and
update tests accordingly.

Reviewed By: dannysu

Differential Revision: D57013627

fbshipit-source-id: d2939a401d508e9a448c554083baab521f71498b