Fix object lowering
Summary:
There is a subtle correctness issue present in lowering currently. If a
numeric property is inserted into the buffer, we continue to use PrStore
on non-numeric properties after that. So, change lowering so that
PrStore is only used *before* numeric props.

This diff also commits numeric props with unserializable values into the
buffer, and patches them with `StoreOwnPropertyInst`s. This was found
to be a speed improvement.

Reviewed By: neildhar

Differential Revision: D57286347

fbshipit-source-id: 721c2a1093e9156674209cd2214d08883dba3457