Disambiguate TestHelpers.h
Summary:
Rename `VMRuntime/TestHelpers.h` to `VMRuntimeTestHelpers.h`. This
naming is in line with other helpers like those for CDP, and makes the
build simpler by not needing to distinguish it from other TestHelpers.h
headers.

There is another `TestHelpers.h` in BCGen, but leave that alone for now
because we aren't enabling BCGen tests.

Reviewed By: avp

Differential Revision: D57452572

fbshipit-source-id: 04b602a6ec0312e526325709d676fed149e0c7ea