Condense inherited classes into StringLiteralTable
Summary:
There were two subclasses of a single mapping class, and the first only
existed to be converted to the other.
This simplifies the types in this file and allows mutation of
`StringLiteralTable` later.

Use `addString` to add strings during `collectStrings()`.
After adding strings, the storage must be populated
(eventually repeatedly, in the case of lazy compilation).

Updated the doc-comments to explain.

NOTE: This diff doesn't change any of the actual functionality or
the behavior to avoid disruption - it's just the interface and number of
classes available to use.
There will be another diff which adds the non-reordering version of the
storage population to make this work properly for lazy compilation.

Reviewed By: tmikov

Differential Revision: D55643721

fbshipit-source-id: 84d5d962d17483defd40b8786c2de310030715c0