Use shared_ptr for Module and SemContext
Summary:
This mostly to remove noise from the next diff: D55699204

We need to store these in `BytecodeProviderFromSrc` for use when lazily
compiling functions in the same `Module`, so use `shared_ptr` for them
just like we do for `Context`.

Doesn't change any semantics or ownership,
just allocates the structures in `shared_ptr` for now.

Reviewed By: tmikov

Differential Revision: D55699182

fbshipit-source-id: fc9a814cd7c3a64e5b1be9311f252182dcf2718b