Use PersistentScopedMap in SemanticResolver
Summary:
We want to be able to persist binding table scopes, so use the
`PersistentScopedMap` for the binding table in `SemanticResolver`.

The only issue here is that `PersistentScopedMap::eraseFromCurrentScope`
didn't exist, so it had to be added here to allow for temporarily adding
the binding for `arguments` in non-simple parameters lists of functions,
but erasing from the linked list could be inefficient.
Instead, just make a new scope in the rare instances where we'd want to
declare `arguments` temporarily, and pop it after visiting the
parameters.

Reviewed By: tmikov

Differential Revision: D54641680

fbshipit-source-id: 407db7d112d00157141cdf058b5b748ab6b51bda