Add ESTreeIRGen for lazy compilation
Summary:
Add a `doLazyFunction` to `ESTreeIRGen`, which will generate IR for lazy
compiled functions.

This has to go back and emit any new declarations for
`UndeclaredGlobal` vars that were introduced by the lazy function.
Other than that, it just reads the data from the
`LazyCompilationDataInst` and runs IRGen until the queue is drained.

Reviewed By: tmikov

Differential Revision: D55832355

fbshipit-source-id: fb5e5d59dbd60308fcb691dc60bd77ac398375b3