Parse let at the start of expression statements
Summary:
We were failing to parse constructions like
`let + 1`
which are valid in loose mode, because we were detecting `let` and
always interpreting it as a declaration.

However, in loose mode, `let` can be an identifier so it has to work in
all expressions including expression statements where it begins a
statement.

Reviewed By: fbmal7

Differential Revision: D56733233

fbshipit-source-id: 83473b2c27a8b46eea4d48aa4c338c0c8851a0e3