Use condition variable to block on background tasks
Summary:
Remove the use of `std::future` and `std::promise` in the Hades
executor implementation, and instead use a condition variable for
synchronisation. This is simpler, and makes it easier to combine this
with `SerialExecutor`.

We have also observed some TSAN reported races in `std::future` during
teardown of the GC involving its shared state with `std::promise`. As
far as I can tell, the current usage is compliant, but it is easier to
reason about the simpler primitives.

Reviewed By: avp

Differential Revision: D57732335

fbshipit-source-id: 83e7a10e5be925df562374df6ca87d7d0612e32b