Write the correct node/edge count in HeapSnapshot
Summary:
Chrome 125 requires the node/edge count to successfully load the heap
snapshot file. Hermes writes the snapshot on the fly, so we can't know
them until we write out the entire nodes/edges array. The workaround we
take is to run two passes: creating a dummy snapshot in the first pass
(without actual I/O) and collecting the node/edge count, then write the
actual snapshot in the second pass.

Reviewed By: neildhar

Differential Revision: D57881365

fbshipit-source-id: e4e380f11afd7b892bd51a4f80157aa7de53b90f