Avoid lazy compilation in formal parameters
Summary:
Capturing functions in parameter initializers are supposed to use the
parameter's environment.
They can't access variables inside the function body, but they can see
previously declared parameters.

However, parameters aren't in a separate scope for the purposes of
resolving variables in the function body - you can redeclare parameters
in the function body and that's supposed to be the same variable.

In eager compilation this just works because we don't add the function
body declarations to the binding table until after running resolution on
default parameter initializers.
In lazy compilation, this doesn't work because we don't run resolution
on the parameter initializers until after they're called, which could
happen well after the binding table has been populated with the
function body's variables.

This happens infrequently enough that we simply eagerly compile all such
functions for now, to avoid reworking the way scopes/redeclaration are
handled completely.

Reviewed By: neildhar

Differential Revision: D58147864

fbshipit-source-id: 4922d5ae617eaba2bc399c4b3ce110238e57d9e0