Add "description" to Flow schema generation
Summary:
@public
This diff adds description parsing to Flow based schema generation. The `description` field already existed in the docblock ir and we were just not setting it. This diff validates the docblock description then adds it to the field.

I also discovered that the location on the comment text is a relative location to the comment (so starts at 0) and always errors at the beginning of the file. I added another field to the comment attachment visitor to track the actual location of the comment.

Reviewed By: tyao1

Differential Revision: D58156907

fbshipit-source-id: c32b8188e09c0c0391b51e7fdc146118037d5f45