Enable runtime on stack when contiguous heap is off
Summary:
Enable putting the runtime on the stack in all builds where the
contiguous heap is disabled.

This diff also simplifies the definition to an explicit define rather
than deciding based on a combination of other defines. This is because
the feature does not depend on any facebook specific code, and in
keeping with us generally moving to more explicit definitions to enable
features rather than preprocessor logic (as we have done with compressed
pointers, the contiguous heap, and boxed doubles).

Reviewed By: lavenzg

Differential Revision: D58449285

fbshipit-source-id: d88bb8b9a113a4058f5b7aa58124c3857935ef75