Don't set value in aliasForOutput
Summary:
`aliasForOutput` should't eagerly write over its input
PinnedHermesValue. When we write to a mutable
handle, we assert that the value it contains is not invalid.

Writing over the input value also limits the usefulness
of this method in the interpreter, since we would be forced
to read all input registers before calling this method.

Reviewed By: avp

Differential Revision: D54924481

fbshipit-source-id: cb10001cea69fe8b27aba2bb606c3f12632c28fc