Fix handling of this argument to HostFunction in TracingRuntime
Summary:
Original Author: neildhar@meta.com
Original Git: 8d6af8994558a24d7f54402c63480894abb61c69
Original Reviewed By: lavenzg
Original Revision: D58048722

The `this` argument being passed to a `HostFunction` should be treated
as a new "def", since this is a value being passed to native for the
first time. We were incorrectly treating it as a use, unlike the other
arguments to the function.

Reviewed By: avp

Differential Revision: D58849676

fbshipit-source-id: 8d1c204a226040276d9e6e2b2e9dc914cd49fb44