Debugger eval compiler pipeline
Summary:
This is the whole pipeline for the local eval which will be used by the
debugger.
Primarily, this consists of slight modifications to each of the
SemanticResolver, IRGen, and backend to allow for generating eval code.

* SemanticResolver needs to use a child `SemContext`
* IRGen needs to modify an existing `Module`
* HBC backend needs to make a new `BytecodeModule`

The IR `Function` memory management is also clarified a bit more:
`~BytecodeFunction` will destroy its `Function` if its has one.
This way we avoid leaking dead `Function`s that can never have more code
compiled under them.

Reviewed By: neildhar

Differential Revision: D56792082

fbshipit-source-id: ae33b835996b703a65cf3d44ae8523cf14993e62