Propagate libraries and include directories of hermesvm in CMake
Summary:
hermesvm takes a private dependency on hermesvm_a, which ensures that
things that link against hermesvm don't also link against the static
libraries that it depends on. However, this erases information about
the include directories that we want to expose.

Additionally, if JSI is a dynamic library, we want it to be added to
the link line of anything that depends on hermesvm, since the JSI
symbols are not going to be available in hermesvm itself.

Reviewed By: avp

Differential Revision: D58770736

fbshipit-source-id: e42ad37faf35327fea85b7ae422708ab94640c4d