Change naming scheme of inner generator
Summary:
The inner generator is not accesible to user code. So this
name does not impact the `.name` property. This change
will make the following function:

```
async function simpleReturn() { ... }
```

Created with the pair of outer and inner generator functions
having the names `anon_0_simpleReturn` and
`anon_0_simpleReturn?inner`, rather than
`?anon_0_?anon_0_simpleAwait`.

Reviewed By: avp

Differential Revision: D57754202

fbshipit-source-id: 487d0317c81ff1eb5ca370d5fbd018c3465242b2