fix non optional JSONBlob in generating cdp api
Summary:
We were only handling `std::optional<JSONBlob>` because this is all we've worked with so far.

This diff adds handling of (non optional) `JSONBlob` and adds `Runtime.inspectRequested` that uses it that would enable the next diff in the stack.

It doesn't change the generated file besides renaming
`TRY_ASSIGN_JSON_BLOB` to `TRY_ASSIGN_OPTIONAL_JSON_BLOB`
and
`assignJsonBlob` to `assignOptionalJsonBlob`.

In the next diff D58869335 we use that API for `Runtime.inspectRequested`

Reviewed By: dannysu

Differential Revision: D58869336

fbshipit-source-id: 001b3332a15ee1a20586f822423fcde90fa5e764